LayoutTests/imported/w3c/web-platform-tests/speculation-rules/speculation-tags/invalid-tags.https_tag-level=rule&type=prefetch-expected.txt


FAIL Sec-Speculation-Tags [rule-based]: integer tag assert_false: expected false got true
FAIL Sec-Speculation-Tags [rule-based]: object tag assert_false: expected false got true
FAIL Sec-Speculation-Tags [rule-based]: null value tag assert_false: expected false got true
FAIL Sec-Speculation-Tags [rule-based]: non-printable character tag assert_false: expected false got true
FAIL Sec-Speculation-Tags [rule-based]: string with non-printable character tag assert_false: expected false got true
PASS Sec-Speculation-Tags [rule-based]: non-ascii string tag
FAIL Sec-Speculation-Tags [rule-based]: 0x19 tag assert_false: expected false got true
FAIL Sec-Speculation-Tags [rule-based]: 0x20 tag assert_equals: expected "\" \"" but got ""
FAIL Sec-Speculation-Tags [rule-based]: 0x7E tag assert_equals: expected "\"~\"" but got "~"
FAIL Sec-Speculation-Tags [rule-based]: 0x7F tag assert_false: expected false got true
